test(auth): share the recording auto-create emitter via tests/adapters
Addresses review feedback: the recording AutoCreateEventEmitter was duplicated
in both new auth_groups tests. Move it to tests/adapters/event.py as the public
RecordingAutoCreateEventEmitter (next to MemoryInfrahubEvent) and import it in
both tests.
