+++ backend/tests/component/auth/auth_groups/test_claim_failure_isolation.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from infrahub import config
-from infrahub.auth.auth_groups.emitter import AutoCreateEventEmitter
 from infrahub.auth.auth_groups.filter import ClaimFilter
 from infrahub.auth.auth_groups.service import AutoCreatedGroupsService
 from infrahub.core.manager import NodeManager
 from infrahub.core.node import Node
 from infrahub.core.protocols import CoreAccount, CoreAccountGroup
+from tests.adapters.event import RecordingAutoCreateEventEmitter
 
 if TYPE_CHECKING:
     from infrahub.core.branch import Branch
@@ -33,24 +33,6 @@ class _FailLookupNodeManager(NodeManager):
         return await super().query(*args, **kwargs)
 
 
-class _RecordingEmitter(AutoCreateEventEmitter):
-    """Adapter emitter that records the audit events emitted during one assignment."""
-
-    def __init__(self) -> None:
-        self.created_groups: list[str] = []
-        self.rejected_claims: list[str] = []
-        self.cap_values: list[int] = []
-
-    async def created(self, *, group: CoreAccountGroup, source_pattern: str) -> None:
-        self.created_groups.append(group.name.value)
-
-    async def claim_rejected(self, *, claim: str) -> None:
-        self.rejected_claims.append(claim)
-
-    async def cap_breached(self, *, cap_value: int, dropped_claims: list[str]) -> None:
-        self.cap_values.append(cap_value)
-
-
 async def test_single_failing_claim_does_not_abort_assignment_and_emits_rejected_event(
     db: InfrahubDatabase,
     default_branch: Branch,
@@ -67,7 +49,7 @@ async def test_single_failing_claim_does_not_abort_assignment_and_emits_rejected
     await account.new(db=db, name="Pat Auto", account_type="User", password="pat-password")
     await account.save(db=db)
 
-    emitter = _RecordingEmitter()
+    emitter = RecordingAutoCreateEventEmitter()
     service = AutoCreatedGroupsService(
         db=db,
         account=account,
